Ignore Ansible retry files
Retry files are merely a by-product of an Ansible run and therefor a
perfect candidate for ignoring within version control.

Ansible retry files are created when a playbook fails and captures the
host(s) that failed to facilitate retrying the playbook against only
those hosts and have no significance beyond that specific environment.