RI-463 Collect OpenStack service logs as artifacts
To assist in debugging Tempest test failures in Jenkins, where
deployments are deleted at the conclusion of a job, save the OpenStack
service logs as artifacts.

Download logs from overcloud nodes to the director, by way of a script
executed on the director, due to lack of SSH access directly from the
host, and then download from the director to the host, by way of
Ansible.

Issue: https://rpc-openstack.atlassian.net/browse/RI-463