ASC-1263: Disable OVN by default
Neither Red Hat nor RPC-R enables OVN in OSP 13 by default, so follow
suite.

While there are benefits to OVN (eg intermediate Linux bridge is not
necessary to implement security groups), there are detriments,
primarily the lack of familiarity and documentation (or at least lack
of emphasis on OVN, and instead ML2/ovs, in documentation), so disable
OVN by default, but allow it to be enabled with an environment
variable.