ASC-1031 Run system tests based on job scenario
This commit parses the `RE_JOB_SCENARIO` to determine additional system
tests to execute based on job context.  The testing scenario is taken
from the last segment of the `RE_JOB_SCENARIO` with any preceding
numerical prefix stripped from the segment.  For example, a value of
`3ctlr_2comp_3ceph` for the `RE_JOB_SCENARIO` will be parsed into the
value of `ceph` for the `SYS_JOB_SCENARIO` variable.  This value is then
appended to the list of system test molecule scenarios to execute for
the system-test job run.

This update allows system test for incompatible scenarios to
introspected and tested in context of the given job. For example, `ceph`
and `swift` scenarios can be run in different deployment jobs and the
undefined scenario will not be run.